+++ src/Plugin/ACFPlugin.php
@@ -10,6 +10,19 @@ class ACFPlugin {
 	public static $acf_folder;
 	private $config;
 
+
+	/**
+	 * Add settings to acf
+	 */
+	public function addSettings()
+	{
+		$acf_settings = $this->config->get('acf', []);
+
+		foreach ($acf_settings as $name=>$value)
+			acf_update_setting($name, $value);
+	}
+
+	
 	public function __construct($config)
 	{
 		$this->config = $config;
@@ -18,11 +31,19 @@ class ACFPlugin {
 
 		add_filter('acf/settings/save_json', function(){ return $this::$acf_folder; });
 		add_filter('acf/settings/load_json', function(){ return [$this::$acf_folder]; });
-		
-		add_filter('acf/fields/google_map/api', function( $api ){
 
-			$api['key'] = $this->config->get('gmap_api_key');
-			return $api;
-		});
+		// When viewing admin
+		if( is_admin() )
+		{
+			add_filter('acf/fields/google_map/api', function( $api ){
+
+				$api['key'] = $this->config->get('gmap_api_key');
+				return $api;
+			});
+
+			// Setup ACF Settings
+			add_action( 'acf/init', [$this, 'addSettings'] );
+
+		}
 	}
 }